Update PutBond method to handle bond updates

+++ Quotations Board Backend/DTOs/Bonds/NewBondDTO.cs
@@ -9,11 +9,11 @@ public class NewBondDTO
     [Required]
     public DateTime MaturityDate { get; set; }
     [Required]
-    public decimal OutstandingValue { get; set; }
+    public double OutstandingValue { get; set; }
     [Required]
     public string CouponType { get; set; } = null!;
     [Required]
-    public decimal CouponRate { get; set; }
+    public double CouponRate { get; set; }
     [Required]
     public string BondType { get; set; } = null!;
     [Required]
